+++ datanode/networkhistory/fsutil/fsutil.go
@@ -17,59 +17,10 @@ package fsutil
 
 import (
 	"archive/zip"
-	"crypto/md5"
-	"encoding/hex"
 	"fmt"
 	"io"
-	"os"
 	"path/filepath"
-
-	vgfs "code.vegaprotocol.io/vega/libs/fs"
 )
-
-// RemoveAllFromDirectoryIfExists used in place of os.RemoveAll when the directory should be emptied but not removed.
-func RemoveAllFromDirectoryIfExists(dir string) error {
-	exists, err := vgfs.PathExists(dir)
-	if err != nil {
-		return err
-	}
-	if !exists {
-		return nil
-	}
-
-	err = filepath.Walk(dir, func(file string, fi os.FileInfo, err error) error {
-		if file != dir {
-			err := os.RemoveAll(file)
-			if err != nil {
-				return fmt.Errorf("failed to remove file:%w", err)
-			}
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		return fmt.Errorf("failed to walk directory:%w", err)
-	}
-
-	return nil
-}
-
-func Md5Hash(path string) (string, error) {
-	hash := md5.New()
-	file, err := os.Open(path)
-	if err != nil {
-		return "", err
-	}
-	defer file.Close()
-
-	_, err = io.Copy(hash, file)
-	if err != nil {
-		return "", err
-	}
-
-	return hex.EncodeToString(hash.Sum(nil)), nil
-}
 
 type readerAtWrapper struct {
 	r io.ReadSeeker